Complete hw6 with email reset and image upload

+++ src/services/contacts.js
@@ -1,3 +1,4 @@
+import mongoose from 'mongoose';
 import { Contact } from '../models/contactModel.js';
 
 export const getAllContacts = async (userId, filter = {}, options = {}) => {
@@ -5,6 +6,7 @@ export const getAllContacts = async (userId, filter = {}, options = {}) => {
 };
 
 export const getContactById = async (contactId, userId) => {
+  if (!mongoose.Types.ObjectId.isValid(contactId)) return null;
   return Contact.findOne({ _id: contactId, userId });
 };
 
@@ -14,6 +16,7 @@ export const addContact = async (data) => {
 };
 
 export const updateContactById = async (contactId, userId, updateData) => {
+  if (!mongoose.Types.ObjectId.isValid(contactId)) return null;
   return Contact.findOneAndUpdate({ _id: contactId, userId }, updateData, {
     new: true,
     runValidators: true,
@@ -21,5 +24,6 @@ export const updateContactById = async (contactId, userId, updateData) => {
 };
 
 export const deleteContactById = async (contactId, userId) => {
+  if (!mongoose.Types.ObjectId.isValid(contactId)) return null;
   return Contact.findOneAndDelete({ _id: contactId, userId });
 };
